requirements.txt
fastapi
uvicorn
pillow
rembg
python-barcode[images]
onnxruntime
python-multipart
cups